pyproject.toml
[project]
name = "robust-dfl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "configspace>=1.2.1",
    "cvxpy>=1.6.0",
    "cvxpylayers>=0.1.8",
    "diffcp>=1.1.3",
    "gurobipy>=12.0.0",
    "ipykernel>=6.29.5",
    "joblib>=1.4.2",
    "jupyter>=1.1.1",
    "lightning==2.4.0",
    "markdown>=3.7",
    "mlflow>=2.18.0",
    "multidict>=6.1.0",
    "nbformat>=5.10.4",
    "networkx>=3.4.2",
    "nptyping>=2.5.0",
    "numpy>=1.26.4",
    "ortools<9.10",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=5.24.1",
    "protobuf>=5.26.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.1",
    "pytorch-lightning==2.4.0",
    "pyyaml>=6.0.2",
    "qpth>=0.0.18",
    "ray>=2.43.0",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "scs>=3.2.7",
    "smac>=2.3.1",
    "tensorboard>=2.18.0",
    "torchmetrics>=1.6.0",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
]
[tool.ruff]
line-length = 100
ignore = [
    "F722"
]
